Updating scripts.

+++ bash/test-install.sh
@@ -2,12 +2,25 @@
 
 if [ -e ./meson.build ]; then
 	rm -fr .build
-	meson setup --reconfigure --wipe -D debug=true .build
-	if [ "${?}" != "0" ]; then
-		exit -1
-	fi
 
-	meson compile -C .build
+	meson setup \
+		--buildtype=plain \
+		--prefix=/usr \
+		--libdir=/usr/lib64 \
+		--libexecdir=/usr/libexec \
+		--bindir=/usr/bin \
+		--sbindir=/usr/sbin \
+		--includedir=/usr/include \
+		--datadir=/usr/share \
+		--mandir=/usr/share/man \
+		--infodir=/usr/share/info \
+		--localedir=/usr/share/locale \
+		--sysconfdir=/etc \
+		--localstatedir=/var \
+		--sharedstatedir=/var/lib \
+		--wrap-mode=nodownload \
+		--auto-features=enabled \
+		.build
 	if [ "${?}" != "0" ]; then
 		exit -1
 	fi
@@ -18,6 +31,14 @@ if [ -e ./meson.build ]; then
 			find src -name *.c | grep -v testprogram >> po/POTFILES.in
 		fi
 		meson compile -C .build $(meson introspect --targets .build | jq -r '.[].name' | grep 'update-po')
+	fi
+
+	meson compile \
+		-C .build \
+		-j 4 \
+		--verbose
+	if [ "${?}" != "0" ]; then
+		exit -1
 	fi
 
 	rm -fr ~/tmp/test-install
